fix: apply feedback updates
