+++ api/src/extention/extension.module.ts
@@ -6,7 +6,9 @@
  * 2. All derivative works must include clear attribution to the original creator and software, Hexastack and Hexabot, in a prominent location (e.g., in the software's "About" section, documentation, and README file).
  */
 
-import { Global, Module } from '@nestjs/common';
+import { Global, Module, OnApplicationBootstrap } from '@nestjs/common';
+
+import { LoggerService } from '@/logger/logger.service';
 
 import { CleanupService } from './cleanup.service';
 
@@ -15,4 +17,17 @@ import { CleanupService } from './cleanup.service';
   providers: [CleanupService],
   exports: [CleanupService],
 })
-export class CleanupModule {}
+export class ExtensionModule implements OnApplicationBootstrap {
+  constructor(
+    private readonly loggerService: LoggerService,
+    private readonly cleanupService: CleanupService,
+  ) {}
+
+  async onApplicationBootstrap() {
+    try {
+      await this.cleanupService.pruneExtensionSettings();
+    } catch (error) {
+      this.loggerService.error('Unable to delete unused settings', error);
+    }
+  }
+}
